+++ src/content/model/ProductNavigatorData.ts
@@ -1,19 +1,21 @@
-import { REST_API_URL } from "../../config/settings";
 import { isInViewport } from "../util/isInViewport";
 import { Event } from "./Event";
-import { AmazonSearchItem } from "./AmazonSearchItem";
+import { AmazonItem } from "./AmazonItem";
 import { Device } from "./Device";
 import { Page } from "./Page";
 import { User } from "./User";
 import { TaskEvent } from "./TaskEvent";
+import { debug } from "../util/debug";
 
 export class ProductNavigatorData {
+    log_level: string;
     device: Device;
     events: any[];
     page: Page;
     user: User;
 
-    constructor() {
+    constructor(log_level: string = "info") {
+        this.log_level = process.env.LOG_LEVEL;
         this.device = new Device();
         this.events = [];
         this.page = new Page();
@@ -23,11 +25,15 @@ export class ProductNavigatorData {
     }
 
     pushEvent(event: Event) {
-        // Checks needed?
+        // TODO: implement check event logic
         this.events.push(event);
         // Whenever a new event is pushed to the datalayer, also send it to backend
         let taskEvent = new TaskEvent(this, event);
-        this.send(taskEvent);
+        debug(
+            () => { console.log(taskEvent) },
+            () => { this.send(taskEvent) },
+            this.log_level
+        );
     }
 
     attachEventsfromSearchResults(searchResults: any[] | NodeListOf<Element>) {
@@ -38,13 +44,13 @@ export class ProductNavigatorData {
                 this.attachViewListener(searchResultElement);
             } else {
                 // Element is in viewport -> directly push view event
-                let item = new AmazonSearchItem(searchResultElement);
+                let item = new AmazonItem(searchResultElement);
                 let event = new Event(item, "view");
                 this.pushEvent(event);
             }
             // Attach click listener to all search elements
             searchResultElement.addEventListener("click", () => {
-                let item = new AmazonSearchItem(searchResultElement);
+                let item = new AmazonItem(searchResultElement);
                 let event = new Event(item, "click");
                 this.pushEvent(event);
             });
@@ -55,7 +61,7 @@ export class ProductNavigatorData {
         $(window).on("resize scroll", () => {
             if (isInViewport(htmlElement) && (!htmlElement.isViewed)) {
                 htmlElement.isViewed = true; // prevents entering this clause multiple times
-                let item = new AmazonSearchItem(htmlElement);
+                let item = new AmazonItem(htmlElement);
                 let event = new Event(item, "view");
                 this.pushEvent(event);
             }
@@ -64,7 +70,7 @@ export class ProductNavigatorData {
 
     send(taskEvent: TaskEvent) {
         $.ajax({
-            url: REST_API_URL,
+            url: process.env.REST_API_URL,
             headers: {
             },
             type: "POST",
@@ -72,5 +78,9 @@ export class ProductNavigatorData {
             contentType: "application/json",
             dataType: "json"
         });
+    }
+
+    isUserFromStudy() {
+        return (this.user.taskID) && (this.user.taskID != process.env.COOKIE_VALUE_MISSING);
     }
 }