+++ webpack.config.js
@@ -1,10 +1,13 @@
 const path = require('path');
+const webpack = require('webpack');
+const CopyPlugin = require('copy-webpack-plugin');
 
 
 module.exports = {
     entry: {
         searchResultsPage: path.resolve(__dirname, "src/content/searchResultsPage.ts"),
-        productDetailPage: path.resolve(__dirname, "src/content/productDetailPage.ts")
+        productDetailPage: path.resolve(__dirname, "src/content/productDetailPage.ts"),
+        serviceWorker: path.resolve(__dirname, "src/background/serviceWorker.ts"),
     },
     output: {
         filename: "[name].js",
@@ -24,5 +27,15 @@ module.exports = {
             { test: /\.css$/i, use: ["style-loader", "css-loader"] }
         ],
     },
-
+    plugins: [
+        new CopyPlugin({
+            patterns: [
+                { from: path.resolve(__dirname, "src/style"), to: path.resolve(__dirname, "dist/style") }
+            ]
+        }),
+        new webpack.ProvidePlugin({
+            $: "jquery",
+            jQuery: "jquery"
+        })
+    ]
 };