+++ webpack.q.js
@@ -1,8 +1,7 @@
 const { merge } = require('webpack-merge');
 const config = require('./webpack.config.js');
 const CopyPlugin = require('copy-webpack-plugin');
-const path = require('path');
-const webpack = require('webpack');
+const Dotenv = require('dotenv-webpack');
 
 
 module.exports = merge(config, {
@@ -11,13 +10,27 @@ module.exports = merge(config, {
     plugins: [
         new CopyPlugin({
             patterns: [
-                { from: "public/staging" },
-                { from: path.resolve(__dirname, "src/style"), to: path.resolve(__dirname, "dist/style") }
+                { from: "public/q" },
+                {
+                    from: "public/manifest.json",
+                    to: "manifest.json",
+                    transform(content, _) {
+                        // https://stackoverflow.com/questions/44232366/how-do-i-build-a-json-file-with-webpack/54700817#54700817
+                        // copy-webpack-plugin passes a buffer
+                        var manifest = JSON.parse(content.toString());
+                        // make any modifications you like, such as
+                        manifest.name = "Product Navigator BETA";
+                        manifest.action.default_title = "Product Navigator BETA";
+                        manifest.description = "Simply remove decision biases from Amazon UK search pages. (TEST VERSION)";
+                        // pretty print to JSON with two spaces
+                        return JSON.stringify(manifest, null, 2);
+                    }
+                }
             ]
         }),
-        new webpack.ProvidePlugin({
-            $: "jquery",
-            jQuery: "jquery"
-        })
+        new Dotenv({
+            path: '.env/.env_q',
+            systemvars: true
+        }),
     ]
 })